feat: subscription

+++ lib/app/data/models/product_model.dart
@@ -137,17 +137,17 @@ class Categories {
 
 class Nutrients {
   int? id;
-  int? serving;
+  double? serving;
   String? unit;
-  int? energy; // Updated to int
-  double? fat; // Updated to double
+  double? energy; 
+  double? fat; 
   double? saturatedFat;
   double? carbohydrate;
   double? sugar;
   double? fiber;
   double? protein;
   double? salt;
-  int? alcohol;
+  double? alcohol;
   String? productId;
 
   Nutrients({
